Delete Contrib folder originally used for DiscUtils
This is the other half of fixing issue #35. The contents of this folder had issues with cloning that would usually fail on Windows. But WSL and MacOS apparently worked fine (aside from certain Tests zip files having incorrect file sizes).

Because DiscUtils is now referenced as NuGet packages, this folder is no longer needed. This also means that users will no longer need to build these projects, and that PSXPrev can be built and run right out of the box.