Switch to Toolkit.Init PreferNative backend (#190)
This solves long-standing issue #40, where a user could not start the program because GLControl would throw an AccessViolationException during the HandleCreated event. Alternatively we can just *not* call Toolkit.Init, since GLControl does that, but it doesn't hurt to call it early if we're calling it with the same options.